requirements.txt
spotipy
pyyaml
click